fix types for nonce size
